requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
discord-protos==0.0.2
discord.py==2.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
multidict==6.1.0
ollama==0.3.3
protobuf==5.28.2
sniffio==1.3.1
yarl==1.12.0
